Fold AssetGraph#createAsset into AssetGraph#add
