+++ lib/AssetGraph.js
@@ -133,12 +133,68 @@ class AssetGraph extends EventEmitter {
      * @param {Asset|String|Object} The asset (or spec) to add
      * @return {Asset[]} The assets instances that were added
      */
-    add(asset, fromUrl, incomingRelation) {
-        if (Array.isArray(asset) || (typeof asset === 'string' && !/^[a-zA-Z-\+]+:/.test(asset) && asset.includes('*'))) {
+    add(assetConfig, fromUrl, incomingRelation) {
+        if (Array.isArray(assetConfig) || (typeof assetConfig === 'string' && !/^[a-zA-Z-\+]+:/.test(assetConfig) && assetConfig.includes('*'))) {
             throw new Error('AssetGraph#add does not accept an array or glob patterns, try AssetGraph#addAll');
         }
-        if (!asset.isAsset) {
-            asset = this.createAsset(asset, fromUrl, incomingRelation);
+        let asset;
+        if (assetConfig.isAsset) {
+            // An already instantiated asset
+            asset = assetConfig;
+        } else {
+            if (typeof assetConfig === 'string') {
+                if (/^[a-zA-Z-\+]+:/.test(assetConfig)) {
+                    assetConfig = { url: assetConfig };
+                } else {
+                    assetConfig = { url: this.resolveUrl(fromUrl || this.root, encodeURI(assetConfig)) };
+                }
+            }
+            if (typeof assetConfig.url === 'string') {
+                if (!/^[a-zA-Z-\+]+:/.test(assetConfig.url)) {
+                    assetConfig.url = this.resolveUrl(fromUrl || this.root, assetConfig.url);
+                }
+                assetConfig.url = assetConfig.url.replace(/\#.*$/, '');
+
+                if (this.canonicalRoot) {
+                    if (assetConfig.url.startsWith(this.canonicalRoot)) {
+                        assetConfig.url = assetConfig.url.replace(this.canonicalRoot, this.root);
+                    }
+                }
+            }
+            if (typeof assetConfig.url === 'string') {
+                if (/^data:/.test(assetConfig.url)) {
+                    let parsedDataUrl = resolveDataUrl(assetConfig.url);
+                    if (parsedDataUrl) {
+                        Object.assign(assetConfig, parsedDataUrl);
+                        delete assetConfig.url;
+                        assetConfig.type = assetConfig.type || this.lookupContentType(assetConfig.contentType);
+                    }
+                } else {
+                    // Check if an asset with that url already exists in the graph,
+                    // and if it does, update it with the information contained
+                    // in assetConfig:
+                    const existingAssets = this.findAssets({url: assetConfig.url});
+                    if (existingAssets.length > 0) {
+                        // If multiple assets share the url, prefer the one that was added last
+                        // (should be customizable?)
+                        // ... or somehow enforce that this situation doesn't arise?
+                        asset = existingAssets[existingAssets.length - 1];
+                        asset.init(assetConfig);
+                    }
+                }
+            }
+            if (!asset) {
+                if (typeof assetConfig.url === 'undefined' && incomingRelation) {
+                    assetConfig.incomingInlineRelation = incomingRelation;
+                }
+                if (assetConfig instanceof this.Asset) {
+                    asset = assetConfig;
+                } else if (assetConfig.type) {
+                    asset = new AssetGraph[assetConfig.type](assetConfig);
+                } else {
+                    asset = new this.Asset(assetConfig);
+                }
+            }
         }
         if (!this.idIndex[asset.id]) {
             this.idIndex[asset.id] = asset;
@@ -174,58 +230,6 @@ class AssetGraph extends EventEmitter {
 
     addAsset(...args) {
         return this.add(...args);
-    }
-
-    createAsset(assetConfig, fromUrl, incomingRelation) {
-        if (typeof assetConfig === 'string') {
-            if (/^[a-zA-Z-\+]+:/.test(assetConfig)) {
-                assetConfig = { url: assetConfig };
-            } else {
-                assetConfig = { url: this.resolveUrl(fromUrl || this.root, encodeURI(assetConfig)) };
-            }
-        }
-        if (typeof assetConfig.url === 'string') {
-            if (!/^[a-zA-Z-\+]+:/.test(assetConfig.url)) {
-                assetConfig.url = this.resolveUrl(fromUrl || this.root, assetConfig.url);
-            }
-            assetConfig.url = assetConfig.url.replace(/\#.*$/, '');
-
-            if (this.canonicalRoot) {
-                if (assetConfig.url.startsWith(this.canonicalRoot)) {
-                    assetConfig.url = assetConfig.url.replace(this.canonicalRoot, this.root);
-                }
-            }
-        }
-        if (typeof assetConfig.url === 'string') {
-            if (/^data:/.test(assetConfig.url)) {
-                let parsedDataUrl = resolveDataUrl(assetConfig.url);
-                if (parsedDataUrl) {
-                    Object.assign(assetConfig, parsedDataUrl);
-                    delete assetConfig.url;
-                    assetConfig.type = assetConfig.type || this.lookupContentType(assetConfig.contentType);
-                }
-            } else {
-                const existingAssets = this.findAssets({url: assetConfig.url});
-                if (existingAssets.length > 0) {
-                    // If multiple assets share the url, prefer the one that was added last
-                    // (should be customizable?)
-                    // ... or somehow enforce that this situation doesn't arise?
-                    const existingAsset = existingAssets[existingAssets.length - 1];
-                    existingAsset.init(assetConfig);
-                    return existingAsset;
-                }
-            }
-        }
-        if (typeof assetConfig.url === 'undefined' && incomingRelation) {
-            assetConfig.incomingInlineRelation = incomingRelation;
-        }
-        if (assetConfig instanceof this.Asset) {
-            return assetConfig;
-        } else if (assetConfig.type) {
-            return new AssetGraph[assetConfig.type](assetConfig);
-        } else {
-            return new this.Asset(assetConfig);
-        }
     }
 
     /**
