+++ testdata/add/relativeUrl/bar.png
