+++ testdata/add/relativeUrl/foo.png
